Separate the styling code to separate files with name style.js and change the component name Relates #58
