+++ client/src/Components/Header/index.js
@@ -1,57 +1,41 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
-import LinkItem from '../LinkItem';
 
-const useStyles = makeStyles(() => ({
-  root: {
-    backgroundColor: '#FFFFFF',
-    borderBottom: '3px solid #1D3D59',
-    height: '80px',
-    justifyContent: 'center',
-    alignItems: 'center',
-    boxSizing: 'border-box',
-    '& *': {
-      boxSizing: 'border-box',
-    },
-  },
-  toolBar: {
-    width: '85%',
-  },
-  logo: {
-    flexGrow: 1,
-    color: '#1D3D59',
-    fontWeight: 'bold',
-  },
-}));
+import LinkItem from '../LinkItem';
+import useStyles from './style';
 
 export default function index() {
   const classes = useStyles();
   return (
-    <>
-      <Container maxWidth="lg">
-        <AppBar position="static" className={classes.root}>
-          <Toolbar className={classes.toolBar}>
-            <Typography variant="h6" className={classes.logo}>
-              House Hunting
-            </Typography>
-            <Typography className={classes.linkItem}>
-              <LinkItem underline="none" text="Houses" linkUrl="/" />
-              <LinkItem underline="none" text="Aboutus" linkUrl="/aboutus" />
-              <LinkItem
-                underline="none"
-                text="Signup"
-                bgFilled
-                linkUrl="/signup"
-              />
-              <LinkItem underline="none" text="Signin" linkUrl="/signin" />
-            </Typography>
-          </Toolbar>
-        </AppBar>
-      </Container>
-    </>
+    <Container maxWidth="lg">
+      <AppBar position="static" className={classes.root}>
+        <Toolbar className={classes.toolBar}>
+          <Typography variant="h6" className={classes.logo}>
+            House Hunting
+          </Typography>
+          <Typography className={classes.linkItem}>
+            <LinkItem underline="none" linkUrl="/">
+              Houses
+            </LinkItem>
+            <LinkItem underline="none" linkUrl="/aboutus">
+              Aboutus
+            </LinkItem>
+            <LinkItem
+              underline="none"
+              registerClass="signupLink"
+              linkUrl="/signup"
+            >
+              Signup
+            </LinkItem>
+            <LinkItem underline="none" linkUrl="/signin">
+              Signin
+            </LinkItem>
+          </Typography>
+        </Toolbar>
+      </AppBar>
+    </Container>
   );
 }
